+++ src/util/opcode/Eqri.java
@@ -5,8 +5,8 @@ import java.util.Arrays;
 public class Eqri extends Opcode {
 
 	@Override
-	public int[] execute(int i1, int i2, int i3, int[] register) {
-		int[] result = Arrays.copyOf(register, register.length);
+	public long[] execute(int i1, int i2, int i3, long[] register) {
+		long[] result = Arrays.copyOf(register, register.length);
 
 		if (register[i1] == i2) {
 			result[i3] = 1;
